// src/types/lead.ts
export type Lead = {
  id?: number
  name: string
  cpf: string
  email: string
  telephone: string
  country: string
  state: string
  city: string
  address: string
  zipcode: string
  merchant: string
  product: string
  provider: string
  service: string
  tracking: string
  providerInfo1: string
  providerInfo2: string
  providerInfo3: string
  productImage?: string
  created_at?: string
}